+++ source/common/api/os_sys_calls_impl.cc
@@ -10,7 +10,7 @@
 namespace Envoy {
 namespace Api {
 
-SysCallIntResult OsSysCallsImpl::bind(int sockfd, const sockaddr* addr, socklen_t addrlen) {
+SysCallIntResult OsSysCallsImpl::bind(int64_t sockfd, const sockaddr* addr, socklen_t addrlen) {
   const int rc = ::bind(sockfd, addr, addrlen);
   return {rc, errno};
 }
@@ -20,42 +20,42 @@ SysCallIntResult OsSysCallsImpl::chmod(const std::string& path, mode_t mode) {
   return {rc, errno};
 }
 
-SysCallIntResult OsSysCallsImpl::ioctl(int sockfd, unsigned long int request, void* argp) {
+SysCallIntResult OsSysCallsImpl::ioctl(int64_t sockfd, unsigned long int request, void* argp) {
   const int rc = ::ioctl(sockfd, request, argp);
   return {rc, errno};
 }
 
-SysCallIntResult OsSysCallsImpl::close(int fd) {
+SysCallIntResult OsSysCallsImpl::close(int64_t fd) {
   const int rc = ::close(fd);
   return {rc, errno};
 }
 
-SysCallSizeResult OsSysCallsImpl::writev(int fd, const iovec* iovec, int num_iovec) {
+SysCallSizeResult OsSysCallsImpl::writev(int64_t fd, const iovec* iovec, int num_iovec) {
   const ssize_t rc = ::writev(fd, iovec, num_iovec);
   return {rc, errno};
 }
 
-SysCallSizeResult OsSysCallsImpl::readv(int fd, const iovec* iovec, int num_iovec) {
+SysCallSizeResult OsSysCallsImpl::readv(int64_t fd, const iovec* iovec, int num_iovec) {
   const ssize_t rc = ::readv(fd, iovec, num_iovec);
   return {rc, errno};
 }
 
-SysCallSizeResult OsSysCallsImpl::recv(int socket, void* buffer, size_t length, int flags) {
+SysCallSizeResult OsSysCallsImpl::recv(int64_t socket, void* buffer, size_t length, int flags) {
   const ssize_t rc = ::recv(socket, buffer, length, flags);
   return {rc, errno};
 }
 
-SysCallSizeResult OsSysCallsImpl::recvmsg(int sockfd, struct msghdr* msg, int flags) {
+SysCallSizeResult OsSysCallsImpl::recvmsg(int64_t sockfd, struct msghdr* msg, int flags) {
   const ssize_t rc = ::recvmsg(sockfd, msg, flags);
   return {rc, errno};
 }
 
-SysCallIntResult OsSysCallsImpl::ftruncate(int fd, off_t length) {
+SysCallIntResult OsSysCallsImpl::ftruncate(int64_t fd, off_t length) {
   const int rc = ::ftruncate(fd, length);
   return {rc, errno};
 }
 
-SysCallPtrResult OsSysCallsImpl::mmap(void* addr, size_t length, int prot, int flags, int fd,
+SysCallPtrResult OsSysCallsImpl::mmap(void* addr, size_t length, int prot, int flags, int64_t fd,
                                       off_t offset) {
   void* rc = ::mmap(addr, length, prot, flags, fd, offset);
   return {rc, errno};
@@ -66,13 +66,13 @@ SysCallIntResult OsSysCallsImpl::stat(const char* pathname, struct stat* buf) {
   return {rc, errno};
 }
 
-SysCallIntResult OsSysCallsImpl::setsockopt(int sockfd, int level, int optname, const void* optval,
-                                            socklen_t optlen) {
+SysCallIntResult OsSysCallsImpl::setsockopt(int64_t sockfd, int level, int optname,
+                                            const void* optval, socklen_t optlen) {
   const int rc = ::setsockopt(sockfd, level, optname, optval, optlen);
   return {rc, errno};
 }
 
-SysCallIntResult OsSysCallsImpl::getsockopt(int sockfd, int level, int optname, void* optval,
+SysCallIntResult OsSysCallsImpl::getsockopt(int64_t sockfd, int level, int optname, void* optval,
                                             socklen_t* optlen) {
   const int rc = ::getsockopt(sockfd, level, optname, optval, optlen);
   return {rc, errno};
@@ -83,12 +83,12 @@ SysCallIntResult OsSysCallsImpl::socket(int domain, int type, int protocol) {
   return {rc, errno};
 }
 
-SysCallSizeResult OsSysCallsImpl::sendmsg(int fd, const msghdr* message, int flags) {
+SysCallSizeResult OsSysCallsImpl::sendmsg(int64_t fd, const msghdr* message, int flags) {
   const int rc = ::sendmsg(fd, message, flags);
   return {rc, errno};
 }
 
-SysCallIntResult OsSysCallsImpl::getsockname(int sockfd, sockaddr* addr, socklen_t* addrlen) {
+SysCallIntResult OsSysCallsImpl::getsockname(int64_t sockfd, sockaddr* addr, socklen_t* addrlen) {
   const int rc = ::getsockname(sockfd, addr, addrlen);
   return {rc, errno};
 }
